requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
crawl4ai
sqlmodel
groq
pyTelegramBotAPI
python-dotenv
lxml